pyproject.toml
[project]
name = "ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docling>=2.91.0",
    "easyocr>=1.7.2",
    "natsort>=8.4.0",
    "onnxruntime>=1.25.1",
    "pillow>=12.2.0",
    "rich>=15.0.0",
    "tqdm>=4.67.3",
    "typer>=0.21.2",
]
[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
